+++ app/models/StreamerEvent.scala
@@ -9,7 +9,7 @@ case class StreamerEvent(
   eventName: String,
   eventDescription: Option[String] = None,
   eventType: String,
-  currentConfidenceAmount: Long,
+  currentConfidenceAmount: Int,
   isActive: Boolean = true,
   startTime: Instant = Instant.now(),
   endTime: Option[Instant] = None,
